[parallel] move content of parallel.defaultimpl to parallel.basic
